Added an api endpoint for a POST method to add an account to the DB
Created a router for /account that will allow users to add their accounts to the DB.
Added the router to the index.js file so that we are able to add it to our app.
We are currently just returning a json message that the account has been created and
my version currently returns the accountId for debugging purposes.