Allowed for storing of sessionId in our backend DB
Created a new helper.js file in our api folder to handle the functionality of setting user sessions. Updated the account.sql file to allow for a new column to be added to store the sessionId. Choose to store sessionId instead of sessionString because sessionString includes the username as well and so it can very in length.