Created /logout route for accounts
Sets the sessionId column to null in our DB and clears the client side sessionString cookie. There is a slight problem as it throws an error if we try to logout with an incorrect username or try and logout multiple times for the same user without logging back in, but we can take care of this on the frontend.