Added an AuthForm that allows the user to login or signup
Created a new Root and Home component. The Root component will render the Home component if the user in logged in, the AuthForm otherwise. This is not currently set up, but will be in the future. The AuthForm is currently updating its state with the users inputs, but is not doing anything with this information currently.