Added a reducer and action for account signup
Created account.js in the actions folder to hold whether the user is currently logged in. Created an account.js file in reducers folder as well to go along with the action. Added the functionality to the AuthForm page to connect to the signup button so that users can fill out the form and actually have accounts created on our backend for them.