Fixed the problem where the user was getting logged out when refreshing the page
We now hit the /authenticated endpoint on the backend in our index.html file, before we render the page. If the user has a valid session then we keep them logged in, otherwise we direct them to the AuthForm component as we have done before.
Changed how we show error messages in the AuthForm component, because when the user does not have a valid session going then it shows an error for the user, but this should not happen when the user has not done anything yet, so we now keep a state to show whether the user has pressed a button yet and if they have not pressed any buttons yet then we hide the errors from the user.