Implemented the displaying of AccountDragons for the user
Created the relevant accountDragon reducer and accountDragon action files.
Created a new AccountDragon component that will show all of the users account dragons and a new AccountDragonRow component that shows the DragonAvatars and nickname for each dragon on its own row.
Updated index.js and Home component so that we are currently showing our accountDragons for the user on the same page as our Dragon component.
Updated the types file so that we had custom states for our AccountDragons.
Updated the account.js file so that we are now exporting fetchFromAccount function and the function was changed so that it now works with our new AccountDragon type as well as Account type.