Updated the AccountDragonRow component to allow for changing of dragon's nicknames
Added a button to each AccountDragonRow that allows the user to switch between Edit and Save mode.
Updated the display of the nickname for each dragon to an input text box that only allows the user to edit it when it is in edit mode.
The state of the component is being stored locally as nothing that is happening here is really effecting the overall redux store.